feat: refactor side-panel
