+++ packages/shared/src/hooks/animation/spring-options.ts
@@ -3,3 +3,5 @@ export type SpringOptions = {
     damping?: number;
     mass?: number;
 };
+
+export type AnimationValues = Record<string, [number | string, number | string]>;
